fix average rating to be float

+++ db/migrate/20150524082853_add_rating_to_review.rb
@@ -1,5 +1,5 @@
 class AddRatingToReview < ActiveRecord::Migration
   def change
-    add_column :reviews, :rating, :integer, default: 1
+    add_column :reviews, :rating, :float, default: 1
   end
 end
